refactor: 移動產品分類的Resource到更好的位置
